Support finding Routes rather than just RouteActions
