+++ cmd/entrypoint/testutil_fake_mapping_cors_test.go
@@ -7,9 +7,6 @@ import (
 	bootstrap "github.com/datawire/ambassador/v2/pkg/api/envoy/config/bootstrap/v3"
 	v3listener "github.com/datawire/ambassador/v2/pkg/api/envoy/config/listener/v3"
 	route "github.com/datawire/ambassador/v2/pkg/api/envoy/config/route/v3"
-	http "github.com/datawire/ambassador/v2/pkg/api/envoy/extensions/filters/network/http_connection_manager/v3"
-	"github.com/datawire/ambassador/v2/pkg/envoy-control-plane/resource/v3"
-	"github.com/datawire/ambassador/v2/pkg/envoy-control-plane/wellknown"
 
 	"github.com/stretchr/testify/assert"
 )
@@ -78,7 +75,8 @@ spec:
 
 	assert.NotNil(t, listener)
 
-	routeAction := findVirtualHostRoute(listener, func(r *route.RouteAction) bool {
+	// Here we're looking for a route whose _action_ is to route to the cluster we want.
+	routeAction := findVirtualHostRouteAction(listener, func(r *route.RouteAction) bool {
 		return r.GetCluster() == "cluster_foo_default_default"
 	})
 	assert.NotNil(t, routeAction)
@@ -88,42 +86,4 @@ spec:
 		assert.Contains(t, []string{"bar.example.com", "foo.example.com"}, m.GetExact())
 
 	}
-}
-
-func findVirtualHostRoute(listener *v3listener.Listener, predicate func(*route.RouteAction) bool) *route.RouteAction {
-	for _, fc := range listener.FilterChains {
-		for _, filter := range fc.Filters {
-			if filter.Name != wellknown.HTTPConnectionManager {
-				continue
-			}
-			hcm := resource.GetHTTPConnectionManager(filter)
-			if hcm != nil {
-				rs, ok := hcm.RouteSpecifier.(*http.HttpConnectionManager_RouteConfig)
-				if ok {
-					for _, vh := range rs.RouteConfig.VirtualHosts {
-						for _, vhr := range vh.Routes {
-							routeAction, ok := vhr.Action.(*route.Route_Route)
-							if ok {
-								if predicate(routeAction.Route) {
-									return routeAction.Route
-								}
-							}
-						}
-					}
-				}
-			}
-		}
-
-	}
-	return nil
-
-}
-
-func findListener(envoyConfig *bootstrap.Bootstrap, predicate func(*v3listener.Listener) bool) *v3listener.Listener {
-	for _, listener := range envoyConfig.StaticResources.Listeners {
-		if predicate(listener) {
-			return listener
-		}
-	}
-	return nil
 }
